+++ packages/pds/src/actor-store/index.ts
@@ -10,7 +10,7 @@ import { RecordReader } from './record/reader'
 import { PreferenceReader } from './preference/reader'
 import { RepoReader } from './repo/reader'
 import { RepoTransactor } from './repo/transactor'
-import { PreferenceTransactor } from './preference/preference'
+import { PreferenceTransactor } from './preference/transactor'
 import { InvalidRequestError } from '@atproto/xrpc-server'
 import { RecordTransactor } from './record/transactor'
 import { CID } from 'multiformats/cid'
